Extract generate_api_struct, clean up docs
